Convert load quantity to double in AutoDesignService

+++ app/src/main/java/com/electricaldesignengineer/app/AutoDesignService.kt
@@ -726,7 +726,7 @@ object AutoDesignService {
 
                 ProfessionalEngineeringCore.LoadInput(
                     name = load.name,
-                    quantity = load.quantity,
+                    quantity = load.quantity.toDouble(),
                     unitPowerKW = load.unitKW,
                     demandFactor = load.demandFactor,
                     powerFactor = load.powerFactor
